Converted the null-render test to the new test scaffolding
